+++ ricaun.RevitTest.Console/Utils/RevitTestUtils.cs
@@ -11,6 +11,9 @@ using System.Threading;
 
 namespace ricaun.RevitTest.Console.Utils
 {
+    /// <summary>
+    /// RevitTestUtils
+    /// </summary>
     public static class RevitTestUtils
     {
         private const int RevitMinVersionReference = 2021;
@@ -27,32 +30,28 @@ namespace ricaun.RevitTest.Console.Utils
             if (RevitUtils.TryGetRevitVersion(filePath, out var revitVersion))
             {
                 Log.WriteLine($"RevitTestUtils: {revitVersion}");
-                if (tests.Length == 0)
+                LoggerTest($"RevitTestUtils: {revitVersion}");
+
+                // Problem with AnavRes.dll / adui22res.dll (version -2020)
+                // Problem with UI (version 2024)
+                var revitVersionMinMax = Math.Min(Math.Max(revitVersion, RevitMinVersionReference), RevitMaxVersionReference);
+                if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersionMinMax, out RevitInstallation revitInstallationMinMax))
                 {
-                    // Problem with AnavRes.dll / adui22res.dll (version -2020)
-                    revitVersion = Math.Min(Math.Max(revitVersion, RevitMinVersionReference), RevitMaxVersionReference);
-                    LoggerTest($"Version {revitVersion}");
-                    if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersion, out RevitInstallation revitInstallation))
-                    {
-                        Log.WriteLine($"RevitTestUtils: {revitInstallation.InstallLocation}");
-                        tests = TestEngine.GetTestFullNames(filePath, revitInstallation.InstallLocation);
-                    }
+                    LoggerTest($"GetTest Version {revitVersionMinMax}");
+                    Log.WriteLine($"RevitTestUtils: {revitInstallationMinMax.InstallLocation}");
+                    tests = TestEngine.GetTestFullNames(filePath, revitInstallationMinMax.InstallLocation);
                 }
-                if (tests.Length == 0)
+                else if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersion, out RevitInstallation revitInstallation))
                 {
-                    // Problem with version 2024
-                    revitVersion = RevitMinVersionReference;
-                    LoggerTest($"Version {revitVersion} << ");
-                    if (RevitInstallationUtils.InstalledRevit.TryGetRevitInstallationGreater(revitVersion, out RevitInstallation revitInstallation))
-                    {
-                        Log.WriteLine($"RevitTestUtils: {revitInstallation.InstallLocation}");
-                        tests = TestEngine.GetTestFullNames(filePath, revitInstallation.InstallLocation);
-                    }
+                    LoggerTest($"GetTest Version {revitVersion}");
+                    Log.WriteLine($"RevitTestUtils: {revitInstallation.InstallLocation}");
+                    tests = TestEngine.GetTestFullNames(filePath, revitInstallation.InstallLocation);
                 }
-                LoggerTest($"Length {tests.Length}");
             }
 
 #if DEBUG
+            LoggerTest($"Length {tests.Length}");
+            LoggerTest($"TestEngine {TestEngine.Version.ToString(3)}");
             if (LoggerTests.Any())
             {
                 LoggerTests.AddRange(tests);
@@ -68,7 +67,7 @@ namespace ricaun.RevitTest.Console.Utils
         [Conditional("DEBUG")]
         private static void LoggerTest(object logger)
         {
-            var loggerTest = $"ricaun.Logger.Tests.Tests(\"{logger}\")";
+            var loggerTest = $"{typeof(RevitTestUtils).FullName}(\"{logger}\")";
             Debug.WriteLine(loggerTest);
             LoggerTests.Add(loggerTest);
         }